IAC/temp/function-source/requirements.txt
functions-framework==3.*
google-cloud-aiplatform
google-generativeai
matplotlib
scikit-learn
tqdm
seaborn
pandas
numpy
ipykernel
promptweaver==0.1.13
streamlit
python-dotenv
SQLAlchemy==2.0.9
functions-framework==3.0.0
google-cloud-secret-manager==2.11.0
psycopg2==2.9.10
dynaconf==3.2.6
cloud-sql-python-connector[pg8000]
pytz